requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9.6
Pillow>=9.5.0
python-decouple>=3.8